+++ lib/my_data_state.dart
@@ -1,4 +1,3 @@
 library my_data_state;
 
 export 'src/data_state.dart';
-export 'src/streams.dart';